requirements-render.txt
# Render-compatible requirements - using Pydantic v1 to avoid Rust compilation
# FastAPI and ASGI server
fastapi==0.95.2
uvicorn[standard]==0.22.0

# Database 
sqlalchemy==2.0.23
aiomysql==0.2.0

# Validation - Pydantic v1 (no Rust compilation required)
pydantic==1.10.15

# Configuration
python-dotenv==1.0.0

# Authentication - minimal dependencies
PyJWT==2.8.0
python-multipart==0.0.6

# Rate limiting
slowapi==0.1.9

# Logging
structlog==23.2.0

# HTTP client - minimal
requests==2.31.0

# Core dependencies
typing-extensions==4.8.0
starlette==0.27.0